Feat : add feature search quote by anime title
